fix: move bulk sms types to bulkSMSCommunicationJourney file
