+++ src/inngest/functions/eventNotification/logic.ts
@@ -1,7 +1,6 @@
 import { differenceInDays, format, startOfDay } from 'date-fns'
 
 import { BULK_SMS_COMMUNICATION_JOURNEY_INNGEST_EVENT_NAME } from '@/inngest/functions/sms/bulkSMSCommunicationJourney'
-import { BulkSMSPayload } from '@/inngest/functions/sms/types'
 import { inngest } from '@/inngest/inngest'
 import { getEvents } from '@/utils/server/builderIO/swcEvents'
 import { prismaClient } from '@/utils/server/prismaClient'
@@ -109,24 +108,22 @@ async function getNotificationInformationForEvents(
       })
     }
 
-    const bulkSMSPayload: BulkSMSPayload = {
-      send: true,
-      messages: [
-        {
-          campaignName: `event-reminder-${event.data.slug}-${event.data.state}-${notificationStrategy}`,
-          smsBody,
-          userWhereInput: {
-            phoneNumber: {
-              in: notifications.map(notification => notification.phoneNumber),
-            },
-          },
-        },
-      ],
-    }
-
     await inngest.send({
       name: BULK_SMS_COMMUNICATION_JOURNEY_INNGEST_EVENT_NAME,
-      data: bulkSMSPayload,
+      data: {
+        send: true,
+        messages: [
+          {
+            campaignName: `event-reminder-${event.data.slug}-${event.data.state}-${notificationStrategy}`,
+            smsBody,
+            userWhereInput: {
+              phoneNumber: {
+                in: notifications.map(notification => notification.phoneNumber),
+              },
+            },
+          },
+        ],
+      },
     })
   }
